+++ compiler/src/main/scala/dxCompiler/Main.scala
@@ -147,6 +147,7 @@ object Main {
       "defaults" -> PathOptionSpec.mustExist,
       "defaultInstanceType" -> StringOptionSpec(),
       "execTree" -> ExecTreeFormatOptionSpec(),
+      "executableCreationParallelism" -> IntOptionSpec.one,
       "extras" -> PathOptionSpec.mustExist,
       "inputs" -> PathOptionSpec.listMustExist,
       "input" -> PathOptionSpec.listMustExist.copy(alias = Some("inputs")),
@@ -356,6 +357,10 @@ object Main {
       }
     }
 
+    val executableCreationParallelism: Int = options
+      .getValue[Int]("executableCreationParallelism")
+      .getOrElse(1)
+
     val defaultScatterChunkSize: Int = options.getValue[Int]("scatterChunkSize") match {
       case None => Constants.JobPerScatterDefault
       case Some(size) =>
@@ -419,6 +424,7 @@ object Main {
             locked,
             if (reorg) Some(true) else None,
             useManifests,
+            executableCreationParallelism,
             instanceTypeSelection,
             baseFileResolver
         )
@@ -533,7 +539,8 @@ object Main {
           translator.complexPathValues,
           instanceTypeSelection,
           defaultInstanceType,
-          fileResolver
+          fileResolver,
+          executableCreationParallelism
       )
       val results = compiler.apply(bundle, project, folder)
       // generate the execution tree if requested
@@ -848,6 +855,9 @@ object Main {
         |      -destination <string>  Full platform path (project:/folder).
         |      -execTree [json,pretty]    
         |                             Print a JSON representation of the workflow.
+        |      -executableCreationParallelism <int>
+        |                             The maximum number of platform executables that dxCompiler can
+        |                             create in parallel, defaults to 1.
         |      -extras <string>       JSON file with extra options (see documentation).
         |      -inputs <string>       JSON file with standard-formatted input values. May be
         |                             specified multiple times. A DNAnexus JSON input file is